requirements.txt
asgiref==3.11.1
channels==4.3.2
channels_redis==4.3.0
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
msgpack==1.1.2
pillow==12.1.1
PyJWT==2.12.1
python-decouple==3.8
redis==7.4.0
sqlparse==0.5.5
